==> notebooks/1.Limpieza_dataset_base.ipynb <==
import pandas as pd
import numpy as np
from pathlib import Path

df = pd.read_csv('../data/raw/f1_dataset_base.csv')
df

df.info()

# Hay que cambiar classified_position de object a numérico

df.describe()

df.groupby('year')['driver_code'].nunique().sort_index()

# Hasta el año 2000 no aparecen todos los pilotos que participan en el campeonato
# Por lo tanto estos años no pueden aparecer como referencia para entrenar al modelo ya que estan incompletos

df['points_real'].nunique()

df['points_real'].value_counts()

# El sistema de puntuación ha ido variando con los años, por lo que para que un resultado sea igual en 2005 que en 2025
# tenemos que estandarizar los puntos, los cuales van a ser nuestro indicador para ordenar la clasificación

df['status'].nunique()

df['status'].value_counts()

df['classified_position'].nunique()

df['classified_position'].value_counts()

df[df['classified_position'] == 'R']

df[df['classified_position'] == 'D']

# Hay que marcar como abandono a todos aquellos que no han terminado la carrera o han sido descalificados por el motivo que sea ya que no se les pueden asignar puntos ya que no tienen posicion final
# y por lo tanto hay que tenerlo en cuenta a la hora de crear las features en un futuro, de otro modo podria dar error si incluimos estas carreras en sus estadisticas

INPUT_FILE = '../data/raw/f1_dataset_base.csv'
OUTPUT_FILE = '../data/processed/f1_clean.csv'

def calcular_puntos_modernos(position):
    puntos_map = {
        1: 25, 2: 18, 3: 15, 4: 12, 5: 10,
        6: 8, 7: 6, 8: 4, 9: 2, 10: 1
    }
    return puntos_map.get(position, 0)

def limpiar_dataset(df):
    df = df.copy()
    
    # 1. FILTRAR DESDE 2005
    df = df[df['year'] >= 2005].copy()
    
    # 2. CONVERTIR POSICIONES A NUMÉRICO
    df['position'] = pd.to_numeric(df['classified_position'], errors='coerce')
    df['position'] = df['position'].fillna(999).astype(int)
    
    # 3. RENOMBRAR COLUMNAS
    df = df.rename(columns={
        'driver_code': 'driver',
        'team_name': 'team'
    })
    
    # 4. RECALCULAR PUNTOS CON SISTEMA MODERNO
    df['puntos_carrera'] = df['position'].apply(calcular_puntos_modernos)
    
    # 5. MARCAR ABANDONOS
    df['es_abandono'] = df['position'] == 999
    
    # 6. SELECCIONAR COLUMNAS FINALES
    columnas_finales = [
        'year', 'round', 'driver', 'team', 
        'position', 'puntos_carrera', 'es_abandono'
    ]
    df = df[columnas_finales]
    
    # 7. ORDENAR
    df = df.sort_values(['year', 'round', 'position']).reset_index(drop=True)
    
    return df


if __name__ == "__main__":
    
    df_raw = pd.read_csv(INPUT_FILE)

    df_clean = limpiar_dataset(df_raw)
    
    Path(OUTPUT_FILE).parent.mkdir(parents=True, exist_ok=True)
    df_clean.to_csv(OUTPUT_FILE, index=False)


df_clean

# Ya esta el dataset listo para calcular el resto de features dinámicas
